This code is ready to be merge
I moved the controllers around in the storyboard we can see the
connections a little bit better. I change the tab label to the closet
from “item” to “Closet” and change the square icon to round icon to
match the others.  I also added the welcome label and sign out button
to the profile page so we can logout of parse and log in another user
if have to.  If that’s all okay with you, please merge this branch.
Thanks!